Fixed ai harvesting

+++ AGEproject/src/Structure.h
@@ -19,7 +19,7 @@ class Structure : public Component
 	{
 		std::string productName;
 		float buildTime;
-		glm::vec3 destination;
+		glm::dvec3 destination;
 		// radius it occupies.
 		float radius;
